Extract superclass for all SCRAM-SHA-* client auth methods
